+++ packages/backend/src/database/entities/savedCharts.ts
@@ -1,4 +1,8 @@
-import { ChartConfig, DBFieldTypes } from '@lightdash/common';
+import {
+    ChartConfig,
+    DBFieldTypes,
+    NumberStyleOrAlias,
+} from '@lightdash/common';
 import { Knex } from 'knex';
 
 export const SavedChartsTableName = 'saved_queries';
@@ -120,6 +124,7 @@ export type DbSavedChartAdditionalMetric = {
     sql?: string;
     hidden?: boolean;
     round?: number;
+    compact?: NumberStyleOrAlias;
     format?: string;
     saved_queries_version_id: number;
 };