+++ packages/frontend/src/components/Home/LatestSavedCharts/index.tsx
@@ -1,7 +1,7 @@
 import { AnchorButton } from '@blueprintjs/core';
 import { subject } from '@casl/ability';
 import { LightdashMode } from '@lightdash/common';
-import { FC } from 'react';
+import { FC, useMemo } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useSavedCharts } from '../../../hooks/useSpaces';
 import { useApp } from '../../../providers/AppProvider';
@@ -19,7 +19,15 @@ const LatestSavedCharts: FC<Props> = ({ projectUuid }) => {
     const isDemo = health.data?.mode === LightdashMode.DEMO;
     const { data: savedCharts = [] } = useSavedCharts(projectUuid);
 
-    const userCanManageCharts = user.data?.ability?.can(
+    const userCanViewCharts = user.data?.ability.can(
+        'view',
+        subject('SavedChart', {
+            organizationUuid: user.data?.organizationUuid,
+            projectUuid,
+        }),
+    );
+
+    const userCanManageCharts = user.data?.ability.can(
         'manage',
         subject('SavedChart', {
             organizationUuid: user.data?.organizationUuid,
@@ -31,11 +39,22 @@ const LatestSavedCharts: FC<Props> = ({ projectUuid }) => {
         history.push(`/projects/${projectUuid}/tables`);
     };
 
+    const featuredCharts = useMemo(() => {
+        return savedCharts
+            .sort((a, b) => {
+                return (
+                    new Date(b.updatedAt).getTime() -
+                    new Date(a.updatedAt).getTime()
+                );
+            })
+            .slice(0, 5);
+    }, [savedCharts]);
+
     return (
         <ResourceList
             resourceIcon="chart"
             resourceType="chart"
-            resourceList={savedCharts}
+            resourceList={featuredCharts}
             enableSorting={false}
             defaultSort={{
                 updatedAt: SortDirection.DESC,
@@ -51,7 +70,7 @@ const LatestSavedCharts: FC<Props> = ({ projectUuid }) => {
                         minimal
                         href="https://docs.lightdash.com/get-started/exploring-data/sharing-insights"
                     />
-                ) : userCanManageCharts && !isDemo ? (
+                ) : userCanViewCharts ? (
                     <LinkButton
                         text={`View all ${savedCharts.length}`}
                         minimal